+++ CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 <!-- next-header -->
 ## [Unreleased] - ReleaseDate  
 
+### Added
+- new argument "interval" for periodic exports
+
+### Changed
+- made file argument obligatory
+
 ## [0.1.0] - 2025-09-10
 
 ### Added